Make quest images optional
