+++ libs/model/src/contest/Contests.projection.ts
@@ -9,7 +9,11 @@ import {
 } from '@hicommonwealth/evm-protocols';
 import { config } from '@hicommonwealth/model';
 import { events } from '@hicommonwealth/schemas';
-import { buildContestLeaderboardUrl, getBaseUrl } from '@hicommonwealth/shared';
+import {
+  buildContestLeaderboardUrl,
+  getBaseUrl,
+  getDefaultContestImage,
+} from '@hicommonwealth/shared';
 import { QueryTypes } from 'sequelize';
 import { models } from '../database';
 import { mustExist } from '../middleware/guards';
@@ -18,7 +22,6 @@ import { getWeightedNumTokens } from '../services/stakeHelper';
 import {
   decodeThreadContentUrl,
   getChainNodeUrl,
-  getDefaultContestImage,
   parseFarcasterContentUrl,
   publishCast,
 } from '../utils';
