+++ libs/model/src/utils/index.ts
@@ -3,7 +3,6 @@ export * from './decodeContent';
 export * from './denormalizedCountUtils';
 export * from './farcasterUtils';
 export * from './generateUnsubscribeLink';
-export * from './getDefaultContestImage';
 export * from './getDelta';
 export * from './makeGetBalancesOptions';
 export * from './parseUserMentions';
