+++ libs/shared/src/index.ts
@@ -1,7 +1,7 @@
 export * from './canvas';
 export * from './constants';
-export * from './defaultAvatar';
 export * from './errors';
+export * from './images';
 export * from './regex';
 export * from './types';
 export * from './utils';
