+++ packages/commonwealth/client/scripts/views/pages/CreateQuest/QuestForm/useQuestForm.ts
@@ -3,6 +3,7 @@ import {
   QuestParticipationLimit,
   QuestParticipationPeriod,
 } from '@hicommonwealth/schemas';
+import { getDefaultContestImage } from '@hicommonwealth/shared';
 import { notifyError, notifySuccess } from 'controllers/app/notifications';
 import { numberNonDecimalGTZeroValidationSchema } from 'helpers/formValidations/common';
 import { calculateRemainingPercentageChangeFractional } from 'helpers/number';
@@ -240,7 +241,7 @@ const useQuestForm = ({ mode, initialValues, questId }: QuestFormProps) => {
       description: values.description.trim(),
       end_date: new Date(values.end_date),
       start_date: new Date(values.start_date),
-      image_url: values.image,
+      image_url: values.image || getDefaultContestImage(),
       ...(values?.community && {
         community_id: values.community.value,
       }),
@@ -299,7 +300,7 @@ const useQuestForm = ({ mode, initialValues, questId }: QuestFormProps) => {
       ...(initialValues.start_date !== values.start_date && {
         start_date: new Date(values.start_date),
       }),
-      image_url: values.image,
+      image_url: values.image || getDefaultContestImage(),
       community_id: values?.community?.value || undefined,
       action_metas: questActionSubForms.map((subForm) => ({
         event_name: subForm.values.action as QuestAction,
